moved upload window to component
